Process payloads by batch (#94)
## Description

The PR introduce a new class `BatchWriter` who allows to group standard call to the document controller. Subsequent calls to the same method will be put in a queue and then executed with a `m*` action instead.

Supported methods:
 - `create`
 - `update`
 - `createOrReplace`
 - `replace`
 - `delete`
 - `get`
 - `exists`

There is almost no change in the existing code of a method to use the batch processing of actions. The `BatchController` can be used instead of the usual document controller for critical actions:

```js
function doSomething (sdk: EmbeddedSDK, batchController: BatchController) {
  // standard call to document.create
  await sdk.document.create(index, collection, document);

  // batch call to document create
  await this.batchController.create(index, collection, document);  
}
```

The batch processing has been applied to the payload processing action.

| Before | After |
|:----:|:----|
| 200 req/s | 800 req/s |