[breaking] Change the `measures` property in API and data model to be more intuitive (#246)
## What does this PR do ?

Change `device-manager/devices:linkAsset` API action. The new format is specified in the type `ApiDeviceLinkAssetRequest`. The `measures` is not a table anymore but an array which each element describe a measure with it's `name` and `type` to avoid confusion between key/value.

```
  action: "linkAsset";

  _id: string;

  refresh?: string;

  assetId: string;

  body?: {
    /**
     * Names of the linked measures
     *
     * Array<{ asset: string, device: string }>
     *
     * @example
     *
     * [
     *   { asset: "externalTemperature", device: "temperature" }
     * ]
     */
    measureNames: Array<{ asset: string; device: string }>;
  };
```

The method to register default devices and array also changed accordingly (`Plugin.registerAsset` and `Plugin.registerDevice`).

The `assets` collection changed as well. `measures` become `measureNames` and have the same format as the parameter in the API action

The `models` collection mappings also changed, the `measures` property now have the same format as the parameter in the API action