[breaking] Remove Measures controller `pushAsset` action (#248)
Remove the action to push directly a measure to an asset.

This does not make sense anymore since all measure have to go through the measure ingestion pipeline starting with the Payload Controller.

Measures who were pushed using this controller have a good chance to be metadata that needed historization. Since this PR, #247, metadata changes are also historized.